Hide 'test only' tools behind the dedicated flag
