+++ mithril-common/src/store/adapter/mod.rs
@@ -1,6 +1,5 @@
 //! Define a generic way to store data with the [Store Adapter][store_adapter::StoreAdapter], with
-//! an adapter ([in memory][MemoryAdapter] and two more for testing ([a stub with one
-//! record][DumbStoreAdapter] and one which [always fails][FailStoreAdapter]).
+//! an adapter [in memory][MemoryAdapter] and another [sqlite][SQLiteAdapter].
 
 mod memory_adapter;
 mod sqlite_adapter;
@@ -10,7 +9,11 @@ pub use memory_adapter::MemoryAdapter;
 pub use sqlite_adapter::{SQLiteAdapter, SQLiteResultIterator};
 pub use store_adapter::*;
 
+#[cfg(any(test, feature = "test_only"))]
 mod dumb_adapter;
+#[cfg(any(test, feature = "test_only"))]
 pub use dumb_adapter::DumbStoreAdapter;
+#[cfg(any(test, feature = "test_only"))]
 mod fail_adapter;
+#[cfg(any(test, feature = "test_only"))]
 pub use fail_adapter::FailStoreAdapter;
